# executor: introduce `Allocator` to manage memory used by executors