--- a/executor/point_get.go
+++ b/executor/point_get.go
@@ -259,5 +259,5 @@ func (e *PointGetExecutor) retTypes() []*types.FieldType {
 }
 
 func (e *PointGetExecutor) newFirstChunk() *chunk.Chunk {
-	return chunk.New(e.retTypes(), 1, 1)
+	return chunk.NewChunkWithAllocator(e.ctx.GetSessionVars().MemoryAllocator, e.retTypes(), 1, 1)
 }