--- a/util/chunk/chunk.go
+++ b/util/chunk/chunk.go
@@ -39,7 +39,7 @@ type Chunk struct {
 	// requiredRows indicates how many rows the parent executor want.
 	requiredRows int
 
-	a        Allocator
+	a Allocator
 }
 
 // Capacity constants.