<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
The first step to fix #9396.
Introduce an `Allocator` to manage memory used by chunk in executors.

### What is changed and how it works?
1. Introduce an interface `Allocator`.
2. Introduce two structs which implement `Allocator`.
3. Add `Release` method to `Chunk`.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Side effects

 - Increased code complexity