@qw4990 2019-06-21
Here is a [post](https://blog.cloudflare.com/recycling-memory-buffers-in-go/) about how to recycle memory buffers in Golang.
We can do the same test to see the effect of this PR on GC.