@qw4990 2019-07-22
Close it because it increases the complexity of memory management and the 4.5% profit is not worth doing that.
We should use some other approach to manage our memory.